THE-674 split REST bucket handlers
